Implemented the riddle prompt functionality from the quest popup menus. Use:  Menu: ?=Question Reminder to get it. Talk to the Merchant to get this puzzle quest and try it.  "blue" is the correct answer to the riddle.

+++ src/client/gui/pawsactivemagicwindow.cpp
@@ -81,7 +81,7 @@ void pawsActiveMagicWindow::HandleMessage( MsgEntry* me )
         }
         case psGUIActiveMagicMessage::Remove:
         {
-            for (int i = 0; i < list->GetRowCount(); i++)
+            for (size_t i = 0; i < list->GetRowCount(); i++)
             {
                 pawsListBoxRow *row = list->GetRow(i);
                 pawsTextBox *name = dynamic_cast<pawsTextBox*>(row->GetColumn(0));
